+++ Sources/DTTableViewManager/DTTableViewManager+Registration.swift
@@ -28,17 +28,25 @@ import DTModelStorage
 
 extension DTTableViewManager
 {
-    /// Registers mapping from model class to `cellClass`.
-    ///
-    /// Method will automatically check for nib with the same name as `cellClass`. If it exists - nib will be registered instead of class.
-    open func register<T:ModelTransfer>(_ cellClass:T.Type,
-                                        handler: @escaping (T, T.ModelType, IndexPath) -> Void = { _, _, _ in },
-                                        mapping: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T: UITableViewCell
+    /// Registers mapping for `cellClass`. Mapping will automatically check for nib with the same name as `cellClass` and register it, if it is found. `UITableViewCell` can also be designed in storyboard.
+    /// - Parameters:
+    ///   - cellClass: UITableViewCell subclass type, conforming to `ModelTransfer` protocol.
+    ///   - handler: configuration closure, that is run when cell is dequeued. Please note, that `handler` is called before `update(with:)` method.
+    ///   - mapping: mapping configuration closure, executed before any registration or dequeue is performed.
+    open func register<Cell:ModelTransfer>(_ cellClass:Cell.Type,
+                                        handler: @escaping (Cell, Cell.ModelType, IndexPath) -> Void = { _, _, _ in },
+                                        mapping: ((ViewModelMapping<Cell, Cell.ModelType>) -> Void)? = nil) where Cell: UITableViewCell
     {
         self.viewFactory.registerCellClass(cellClass, handler: handler, mapping: mapping)
     }
     
-    open func register<T: UITableViewCell, U>(_ cellClass: T.Type, for modelType: U.Type, handler: @escaping (T, U, IndexPath) -> Void, mapping: ((ViewModelMapping<T, U>) -> Void)? = nil) {
+    /// Registers mapping from `modelType` to `cellClass`. Mapping will automatically check for nib with the same name as `cellClass` and register it, if it is found. `UITableViewCell` can also be designed in storyboard.
+    /// - Parameters:
+    ///   - cellClass: UITableViewCell to register
+    ///   - modelType: Model type, which is mapped to `cellClass`.
+    ///   - handler: configuration closure, that is run when cell is dequeued.
+    ///   - mapping: mapping configuration closure, executed before any registration or dequeue is performed.
+    open func register<Cell: UITableViewCell, Model>(_ cellClass: Cell.Type, for modelType: Model.Type, handler: @escaping (Cell, Model, IndexPath) -> Void, mapping: ((ViewModelMapping<Cell, Model>) -> Void)? = nil) {
         viewFactory.registerCellClass(cellClass, modelType, handler: handler, mapping: mapping)
     }
     
@@ -46,72 +54,74 @@ extension DTTableViewManager
     ///
     /// Method will automatically check for nib with the same name as `headerClass`. If it exists - nib will be registered instead of class.
     /// This method also sets TableViewConfiguration.sectionHeaderStyle property to .view.
-    /// - Note: Views does not need to be `UITableViewHeaderFooterView`, if it's a `UIView` subclass, it also will be created from XIB.
+    /// - Note: Views does not need to be `UITableViewHeaderFooterView`, if it's a `UIView` subclass, it also will be created from XIB. In the latter case, events defined inside mapping closure are not supported.
+    /// - Note: `handler` closure is called before `update(with:)` method.
     /// - SeeAlso: `UIView+XibLoading`.
-    open func registerHeader<T:ModelTransfer>(_ headerClass : T.Type,
-                                              handler: @escaping (T, T.ModelType, Int) -> Void = { _, _, _ in },
-                                              mapping: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T: UIView
+    open func registerHeader<View:ModelTransfer>(_ headerClass : View.Type,
+                                              handler: @escaping (View, View.ModelType, Int) -> Void = { _, _, _ in },
+                                              mapping: ((ViewModelMapping<View, View.ModelType>) -> Void)? = nil) where View: UIView
     {
         configuration.sectionHeaderStyle = .view
-        viewFactory.registerSupplementaryClass(T.self, ofKind: DTTableViewElementSectionHeader, handler: handler, mapping: mapping)
+        viewFactory.registerSupplementaryClass(View.self, ofKind: DTTableViewElementSectionHeader, handler: handler, mapping: mapping)
     }
     
     /// Registers mapping from model class to header view of `headerClass` type.
     ///
     /// Method will automatically check for nib with the same name as `headerClass`. If it exists - nib will be registered instead of class.
     /// This method also sets TableViewConfiguration.sectionHeaderStyle property to .view.
-    /// - Note: Views does not need to be `UITableViewHeaderFooterView`, if it's a `UIView` subclass, it also will be created from XIB.
+    /// - Note: Views does not need to be `UITableViewHeaderFooterView`, if it's a `UIView` subclass, it also will be created from XIB. In the latter case, events defined inside mapping closure are not supported.
     /// - SeeAlso: `UIView+XibLoading`.
-    open func registerHeader<T: UIView, U>(_ headerClass : T.Type,
-                                           for: U.Type,
-                                              handler: @escaping (T, U, Int) -> Void,
-                                              mapping: ((ViewModelMapping<T, U>) -> Void)? = nil)
+    open func registerHeader<View: UIView, Model>(_ headerClass : View.Type,
+                                                  for: Model.Type,
+                                                  handler: @escaping (View, Model, Int) -> Void,
+                                                  mapping: ((ViewModelMapping<View, Model>) -> Void)? = nil)
     {
         configuration.sectionHeaderStyle = .view
-        viewFactory.registerSupplementaryClass(T.self, ofKind: DTTableViewElementSectionHeader, handler: handler, mapping: mapping)
+        viewFactory.registerSupplementaryClass(View.self, ofKind: DTTableViewElementSectionHeader, handler: handler, mapping: mapping)
     }
     
     /// Registers mapping from model class to footerView view of `footerClass` type.
     ///
     /// Method will automatically check for nib with the same name as `footerClass`. If it exists - nib will be registered instead of class.
     /// This method also sets TableViewConfiguration.sectionFooterStyle property to .view.
-    /// - Note: Views does not need to be `UITableViewHeaderFooterView`, if it's a `UIView` subclass, it also will be created from XIB.
+    /// - Note: Views does not need to be `UITableViewHeaderFooterView`, if it's a `UIView` subclass, it also will be created from XIB. In the latter case, events defined inside mapping closure are not supported.
+    /// - Note: `handler` closure is called before `update(with:)` method.
     /// - SeeAlso: `UIView+XibLoading`.
-    open func registerFooter<T:ModelTransfer>(_ footerClass: T.Type,
-                                              handler: @escaping (T, T.ModelType, Int) -> Void = { _, _, _ in },
-                                              mapping: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T:UIView
+    open func registerFooter<View:ModelTransfer>(_ footerClass: View.Type,
+                                              handler: @escaping (View, View.ModelType, Int) -> Void = { _, _, _ in },
+                                              mapping: ((ViewModelMapping<View, View.ModelType>) -> Void)? = nil) where View:UIView
     {
         configuration.sectionFooterStyle = .view
-        viewFactory.registerSupplementaryClass(T.self, ofKind: DTTableViewElementSectionFooter, handler: handler, mapping: mapping)
+        viewFactory.registerSupplementaryClass(View.self, ofKind: DTTableViewElementSectionFooter, handler: handler, mapping: mapping)
     }
     
     /// Registers mapping from model class to footer view of `footerClass` type.
     ///
     /// Method will automatically check for nib with the same name as `footerClass`. If it exists - nib will be registered instead of class.
     /// This method also sets TableViewConfiguration.sectionFooterStyle property to .view.
-    /// - Note: Views does not need to be `UITableViewHeaderFooterView`, if it's a `UIView` subclass, it will be created from XIB.
+    /// - Note: Views does not need to be `UITableViewHeaderFooterView`, if it's a `UIView` subclass, it will be created from XIB. In the latter case, events defined inside mapping closure are not supported.
     /// - SeeAlso: `UIView+XibLoading`.
-    open func registerFooter<T: UIView, U>(_ footerClass : T.Type,
-                                           for: U.Type,
-                                           handler: @escaping (T, U, Int) -> Void,
-                                           mapping: ((ViewModelMapping<T, U>) -> Void)? = nil)
+    open func registerFooter<View: UIView, Model>(_ footerClass : View.Type,
+                                           for: Model.Type,
+                                           handler: @escaping (View, Model, Int) -> Void,
+                                           mapping: ((ViewModelMapping<View, Model>) -> Void)? = nil)
     {
         configuration.sectionFooterStyle = .view
-        viewFactory.registerSupplementaryClass(T.self, ofKind: DTTableViewElementSectionFooter, handler: handler, mapping: mapping)
+        viewFactory.registerSupplementaryClass(View.self, ofKind: DTTableViewElementSectionFooter, handler: handler, mapping: mapping)
     }
     
     /// Unregisters `cellClass` from `DTTableViewManager` and `UITableView`.
-    open func unregister<T:ModelTransfer>(_ cellClass: T.Type) where T:UITableViewCell {
-        viewFactory.unregisterCellClass(T.self)
+    open func unregister<Cell:ModelTransfer>(_ cellClass: Cell.Type) where Cell:UITableViewCell {
+        viewFactory.unregisterCellClass(Cell.self)
     }
     
     /// Unregisters `headerClass` from `DTTableViewManager` and `UITableView`.
-    open func unregisterHeader<T:ModelTransfer>(_ headerClass: T.Type) where T: UIView {
-        viewFactory.unregisterHeaderClass(T.self)
+    open func unregisterHeader<View:ModelTransfer>(_ headerClass: View.Type) where View: UIView {
+        viewFactory.unregisterHeaderClass(View.self)
     }
     
     /// Unregisters `footerClass` from `DTTableViewManager` and `UITableView`.
-    open func unregisterFooter<T:ModelTransfer>(_ footerClass: T.Type) where T: UIView {
-        viewFactory.unregisterFooterClass(T.self)
+    open func unregisterFooter<View:ModelTransfer>(_ footerClass: View.Type) where View: UIView {
+        viewFactory.unregisterFooterClass(View.self)
     }
 }