+++ Sources/Tests/MappingConditionsTestCase.swift
@@ -12,7 +12,6 @@ import DTTableViewManager
 class MappingConditionsTestCase: BaseTestCase {
     
     func testMappingCanBeSwitchedBetweenSections() {
-        controller.manager.memoryStorage.defersDatasourceUpdates = true
         controller.manager.register(NibCell.self) { mapping in
             mapping.condition = .section(0)
         }
@@ -32,7 +31,6 @@ class MappingConditionsTestCase: BaseTestCase {
     }
     
     func testCustomMappingIsRevolvableForTheSameModel() {
-        controller.manager.memoryStorage.defersDatasourceUpdates = true
         controller.manager.register(NibCell.self) { mapping in
             mapping.condition = .custom({ indexPath, model in
                 guard let model = model as? Int else { return false }
@@ -56,7 +54,6 @@ class MappingConditionsTestCase: BaseTestCase {
     }
     
     func testMappingCanBeSwitchedForNibNames() {
-        controller.manager.memoryStorage.defersDatasourceUpdates = true
         controller.manager.register(NibCell.self) { mapping in
             mapping.condition = .section(0)
             mapping.reuseIdentifier = "NibCell One"